Added sample: settings_mcumgr
This sample was copied verbatim from the SDK repo
(from samples/settings).

The SDK will continue to have a settings sample, but it will be
based on the shell instead of MCUmgr. However, the MCUmgr-based
settings sample is still useful, so that's why it's being copied here.

The reason the SDK is removing MCUmgr from the settings sample is
because it is not currently supported on esp32 when used in combination
with WiFi.

Signed-off-by: Nick Miller <nick@golioth.io>